+++ src/effects/spells/divineFavor.js
@@ -7,6 +7,7 @@ export function divineFavorEffect(document) {
     { key: "system.bonuses.rwak.damage", value: "1d4[Radiant]", mode: CONST.ACTIVE_EFFECT_MODES.CUSTOM, priority: 0 }
   );
   document.effects.push(effect);
+  document.system.damage = { parts: [], versatile: "", value: "" };
 
   return document;
 }