+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT UP
 
 - Better handling of class identifiers for scaling values if they have non-alpha characters.
+- Prevent error when updating character as a player without filesystem browse permissions.
 
 # 3.7.14