+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# NEXT UP
+# 3.7.15
 
 - Better handling of class identifiers for scaling values if they have non-alpha characters.
 - Prevent error when updating character as a player without filesystem browse permissions.